// src/api/generated/urls/urls.ts
export type listAllResponse200 = {
  data: UrlResponse[]
  status: 200
}

export type listAllResponse401 = {
  data: ErrorResponse
  status: 401
}

export type listAllResponseSuccess = listAllResponse200 & {
  headers: Headers
}
export type listAllResponseError = listAllResponse401 & {
  headers: Headers
}

export type listAllResponse = listAllResponseSuccess | listAllResponseError